% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heads and Tails.R
\name{ht}
\alias{ht}
\title{Heads and Tails}
\usage{
ht(x)
}
\arguments{
\item{x}{dater}
}
\value{
first and last 5 lines of dater
}
\description{
displays first and last 5 lines of dater
}
