% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Install and load packages.R
\name{ipak}
\alias{ipak}
\title{Install and load packages}
\usage{
ipak(pkg)
}
\arguments{
\item{pkg}{a vector of packages}
}
\value{
Packages will be installed and loaded.
}
\description{
This will check to see if packages are installed. Install them if they are
not, then load them into the R session.
}
\examples{
## usage
packages <-  c("curl", "httr","plyr","devtools")
ipak(packages)
}
