% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method of moments.R
\name{mom}
\alias{mom}
\title{Method of Moments}
\usage{
mom(distribution, mean, sd)
}
\arguments{
\item{distribution}{enter "beta" or "gamma"}

\item{mean}{mean value of samples}

\item{sd}{standard deviation of samples}
}
\value{
Returns parameters \code{a} (i.e., shape 1) and \code{b} (i.e., shape
  2) of the \strong{beta} distribution.

  Returns parameters \code{c} (i.e., shape) and \code{b} (i.e., rate) of the
  \strong{gamma} distribution.  Note that the scale parameter = 1/b (See example).

  A warning will show if negative values are obtained. distribution.
}
\description{
Method of Moments to determine parameters for gamma or beta distributions.
}
\examples{
##Beta Distribution
##assign beta parms from mean = 0.5 and sd = 0.2 to outz
outz = mom("beta",0.5, 0.2)

##use outz to generate 10 random beta observations from mean = 0.5 and sd = 0.2
rbeta(10,outz[1],outz[2])

##Example of negative parameter estimate function
mom("beta",50, 0.2)

##Gamma Distribution
##Generate some probability levels
prob_levels<-c(0.001,.05,.25,.5,.75,.95,0.999)

##Generate gamma parms with man of 10 and sd of 5
outz <- mom("gamma",10, 5)

##Use the gamma parms (outz) and the probability with qgamma to find the
##quantile values of a gamma dist. Note the 1/b component to calculate the scale parameter
qgamma(prob_levels,outz[1],1/outz[2])
}
