% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OW Tornado diagram.R
\name{OW.tornado}
\alias{OW.tornado}
\title{Tornado Diagram for One-Way Sensitivity Analysis}
\usage{
OW.tornado(data = NULL, best.outcome = max, x.axis.title = "Utility",
  xlim.big = NULL, xlim.small = NULL)
}
\arguments{
\item{data}{dater frame or matrix containing decision, utility, parm
(parameter), pct (increment) as column names.}

\item{best.outcome}{either max or min.  Represents whether the utility is
optimized at minimum or maximun values.}

\item{x.axis.title}{self explanitory}

\item{xlim.big}{optional input to fine tine the upper limit of the x axais in
it gets weird}

\item{xlim.small}{similar to xlim.big}
}
\value{
Returns a barplot in the form of a tornado diagram
}
\description{
Output is a tornado diagram for SDM sensativity analysis.
}
\examples{
##Lets make some fake dater called infile
decision = seq(1,2,1)
parm = paste("parm",seq(1,25,1))
pct = seq(0.05,0.95,length.out = 10)
utility = sample(rnorm(1,0,50):runif(1,50,500),500,replace = T)
infile = cbind(expand.grid(decision=decision,parm = parm,pct = pct),utility = utility)

##Open seperate plotting window and make da plot
x11()
OW.tornado(infile,best.outcome = max,x.axis.title = "Coach Ditka's rating on Madden 1985")

##Save plot to directory
png(file="~/one way sensativity output.png",width = 1000,height = 800,res = 100)
OW.tornado(infile,best.outcome = max,x.axis.title = "Coach Ditka's rating on Madden 1985")
dev.off()
}
