% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round perty.R
\name{round.perty}
\alias{round.perty}
\title{Round Perty}
\usage{
\method{round}{perty}(x, lower_bound = TRUE, nice_small = c(0, 5, 10),
  nice_big = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
}
\arguments{
\item{x}{a number or vector}

\item{lower_bound}{logical to round down(default) or round up}

\item{nice_small}{a vector scalar for rounding small numbers (<100)}

\item{nice_big}{a vector scalar for rounding big numbers (>100)}
}
\value{
Returns rounded number or vector
}
\description{
Rounds a number of vector of numbers to a perty number for plot axis or
whatever else.
}
\examples{
round.perty(c(4,12,233,49488), lower_bound = T)
round.perty(c(4,12,233,49488), lower_bound = F)
}
