% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensativity analysis.R
\name{sensy}
\alias{sensy}
\title{Sensativity analysis function}
\usage{
sensy(vary = NULL, pct = 0.5, dist = "beta", cv = 0.5)
}
\arguments{
\item{vary}{a character string of the parameter name to vary}

\item{pct}{a quantile location (e.g., 0.05,0.5,0.95)}

\item{dist}{"beta" or "gamma"}

\item{cv}{coefficent of variation}
}
\value{
Returns the new value of the vary parm
}
\description{
Outputs a parameter value of the beta or gamma distribution as a function of a mean value, distribution, quantile, and coeficent of variation
}
